% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllDDMSheets.R
\name{list2DDM}
\alias{list2DDM}
\title{Converts a list of data frames into a CompadreDDM}
\usage{
list2DDM(parameters)
}
\arguments{
\item{parameters}{A list produced by \code{readAllSheets}.}
}
\value{
an object of class \code{CompadreDDM}.
}
\description{
Converts a list of data frames into a CompadreDDM
}
