% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop7str.R
\docType{data}
\name{pop7str}
\alias{pop7str}
\title{Population Structure Data from China's 7th Census}
\format{
A tibble with 21 rows and 3 columns:
\describe{
\item{Age}{Age group, character string.}
\item{Male}{Male population count (in millions), numeric.}
\item{Female}{Female population count (in millions), numeric.}
}
}
\source{
7th National Population Census of China.
}
\usage{
pop7str
}
\description{
A dataset containing population figures from the 7th National Population Census of China,
broken down by age group and gender.
}
\examples{
data(pop7str)
head(pop7str)
}
\keyword{datasets}
