% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllDDMSheets.R
\name{readAllSheets}
\alias{readAllSheets}
\title{Reads all the sheets in Compadre Density-dependent matrix Excel sheet}
\usage{
readAllSheets(file)
}
\arguments{
\item{file}{the file path to the Excel sheet.}
}
\value{
A list of data frames
}
\description{
this function reads all the pages in an Excel sheet
and returns a list of data frames. The length of the list is equal to the
number of the sheets in the file. Unfortunately, sheet names are not
preserved so you'll need to insert those manually if you want to use them.
}
