% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projMSM_helpers.R
\name{standardA}
\alias{standardA}
\title{Standardize a single or set of matrices}
\usage{
standardA(A)
}
\arguments{
\item{A}{A single square matrix or array of square matrices stacked as a
3-dimensional array}
}
\value{
An array of the same dimension as the input (\code{A}). They are
the input matrices that have been divided by their dominant eigenvalues.
}
\description{
Standardizes matrices in an array by their dominant eigenvalues. This filters
out the asymptotic components of each matrix, leaving only the transient
growth components.
}
\references{
Add something here
}
