% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompadreMatClasses.R
\docType{class}
\name{Stochastic-Matrix-Classes}
\alias{Stochastic-Matrix-Classes}
\alias{CompadreDDM}
\alias{CompadreMSM}
\alias{CompadreDDM-class}
\alias{CompadreMSM-class}
\alias{CompadreESM-class}
\alias{CompadreESM}
\alias{matrixExpr}
\alias{matrixExpr,CompadreDDM-method}
\alias{matrixExpr,CompadreESM-method}
\alias{matFuns}
\alias{matFuns,CompadreDDM-method}
\alias{matFuns,CompadreESM-method}
\title{Stochastic Matrix Classes}
\usage{
matrixExpr(object)

\S4method{matrixExpr}{CompadreDDM}(object)

\S4method{matrixExpr}{CompadreESM}(object)

matFuns(object)

\S4method{matFuns}{CompadreDDM}(object)

\S4method{matFuns}{CompadreESM}(object)
}
\description{
Stochastic matrix classes in \code{Rcompadre} and \code{popdemo} can be broken
down into two types: stochastic by matrix selection and stochastic by element
selection. Within the latter category, there are three different types supported:
density dependent, environmentally dependent, and their combination (note that
only density dependence is implemented currently, but the last two are coming
in future versions). 

Stochastic by matrix selection matrices are of class \code{CompadreMSM} and
only contain a list of matrices filled with actual numbers. Additional information
for each one should be supplied in call to \code{project}.

Stochastic by element selection matrices are comprised of two lists - one for
constants and the population vector (the \code{dataList}), and one for 
expressions that calculate vital rates based on the population vector, as well
as the expression for the matrix itself (the \code{matExprs} and 
\code{matrixExpr}, respectively). The population vector is optional and can be
set to \code{NA}. Options for stage-biased vectors and random draws from a
dirichlet distribution are provided in \code{project}.
 
Entries in \code{matExprs} that are not the \code{matrixExpression} should read
like R code. If a value in one expression is a function of other constants, 
\code{makeMatExprs} will recognize this and generate the correct order for
evaluating the expressions. For example, fecundity of the second stage 
class could density dependent and defined as \deqn{f_2 = e^{bf_0 + bf_1 * u}}
This should be translated into an expression as \code{f_2 = exp(bf_0 + bf_1 * u)}
(A poisson model with slope \code{bf_1} and intercept \code{bf_0}). \code{u} is
the sum of all non-dormant stages in the population vector. Thus, if you supply
and additional expression, \code{u = sum(stage1, stage2, stage3)}, then the 
expression for \code{u} will get evaluated first and substituted into the 
expression for \code{s_2}. More complicated expressions (provided they are valid
R code) can also be supplied as long as they return a scalar numeric value. 
You do not need to specify these in any particular 
order in calls to \code{makeMatExprs}. See the vignette on writing \code{matExprs}
functions for additional examples (\code{browseVignettes('popdemo')}).
 
The matrix expression can be filled with a
mixture of symbols (e.g. \code{s_2 * f}) and numeric values, provided that
the symbols are defined in either \code{matExprs} or named in the 
\code{dataList}. See the examples in \code{\link{makeMatExprs}}.
}
\section{Slots}{

\describe{
\item{\code{data_list}}{list of named constant values for substitution in matrix. For 
additional examples, see \code{\link{makeDataList}}.}

\item{\code{matExprs}}{list of expressions that define the density dependent components.
This can be constructed by hand, but it is safer to construct this using 
\code{makeMatExprs}.
For additional examples, see \code{\link{makeMatExprs}}.}

\item{\code{A}}{a list of matrices}

\item{\code{env_exprs}}{additional expressions specifying environmental effects}
}}

