% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3plot-projection.R
\name{plot.projection}
\alias{plot.projection}
\title{Plot population dynamics}
\usage{
\method{plot}{projection}(x, labs = TRUE, plottype = "lines",
  ybreaks = 100, shadelevels = 100, show.bounds = FALSE, ...)
}
\arguments{
\item{x}{an object of class 'projection' created using \code{\link{project}}.}

\item{labs}{logical: if \code{TRUE}, then lines are automatically labelled 
when \code{x} contains stage-biased projections (generated using 
\code{vector="n"} in \code{project})}

\item{plottype}{for projections generated from dirichlet draws (see 
\code{\link{project}}), \code{plottype} has two options. \code{"lines"}
will plot each projection as a separate line. \code{"shady"} will 
plot shaded contours showing the probabilities of population 
densities over over time, calculated across the set of projections from 
dirichlet draws. Darker colours indicate more likely population 
densities.}

\item{ybreaks}{if \code{plottype="shady"}, gives the number of breaks on
the y axis for generating the grid for the contour plot. A larger number of
breaks means a finer resolution grid for the contour plot.}

\item{shadelevels}{if \code{plottype="shady"}, gives the number of
colour/shading levels to use when generating the contour plot. A larger
number of levels means a finer resolution colour/shade on the contour
plot of population density.}

\item{show.bounds}{logical: if \code{vector="diri"}, indicates whether to 
plot the bounds on population density.}

\item{...}{arguments to be passed to methods: see \code{\link{par}} and
\code{\link{plot}}.}
}
\description{
Plot dynamics of a population matrix projection model
}
\details{
Plots population dynamics (time series of density) for objects of class
'projection' created using \code{\link{project}}.  The method is
particularly useful for sets of more than one projection.
}
\examples{
  # Create a 3x3 PPM
  ( A <- matrix(c(0,1,2,0.5,0.1,0,0,0.6,0.6), byrow=TRUE, ncol=3) )

  # Create an initial stage structure
  ( initial <- c(1,3,2) )

  # plot stage-biased dynamics of A over 70 intervals using
  # standardised dynamics and log y axis
  plot(project(A,time=70,standard.A=TRUE,standard.vec=TRUE),
       log="y",ylab="Years",xlab="Density")

  # plot a projection of a specified initial stage structure
  # for 10 intervals
  plot(project(A,vector=initial,time=50))

  # Load the desert Tortoise matrix
  data(Tort)

  # Project 500 population vectors from a uniform dirichlet 
  # distribution, and plot the density of population sizes
  # within the bounds of population density
  plot(project(Tort, time=30, vector="diri", draws=500, alpha.draws="unif",
               standard.A=TRUE),plottype="shady", show.bounds=TRUE)

}
\seealso{
\code{\link{project}}
}
\concept{
population projection
}
